fix(landing): restore static export for GitHub Pages deploy
The aoagents swap dropped output: "export", so deploy-landing built
successfully but never wrote out/, and upload-pages-artifact failed.
