+++ frontend/src/landing/next.config.ts
@@ -1,55 +1,25 @@
 import type { NextConfig } from "next";
 
+// GitHub Pages serves a static export (see .github/workflows/deploy-landing.yml).
+// Features that need a Node server (rewrites, proxy.ts) are intentionally omitted.
 const config: NextConfig = {
-  reactStrictMode: true,
-
-  images: {
-    qualities: [75, 80],
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "*.public.blob.vercel-storage.com",
-      },
-      {
-        protocol: "https",
-        hostname: "unavatar.io",
-      },
-    ],
-  },
-
-  async rewrites() {
-    return [
-      {
-        source: "/ingest/static/:path*",
-        destination: "https://us-assets.i.posthog.com/static/:path*",
-      },
-      {
-        source: "/ingest/:path*",
-        destination: "https://us.i.posthog.com/:path*",
-      },
-      {
-        source: "/ingest/decide",
-        destination: "https://us.i.posthog.com/decide",
-      },
-    ];
-  },
-
-  async redirects() {
-    return [
-      {
-        source: "/about",
-        destination: "/team",
-        permanent: true,
-      },
-      {
-        source: "/docs/:path*",
-        destination: "https://docs.superset.sh/:path*",
-        permanent: false,
-      },
-    ];
-  },
-
-  skipTrailingSlashRedirect: true,
+	output: "export",
+	reactStrictMode: true,
+	trailingSlash: true,
+	images: {
+		unoptimized: true,
+		qualities: [75, 80],
+		remotePatterns: [
+			{
+				protocol: "https",
+				hostname: "*.public.blob.vercel-storage.com",
+			},
+			{
+				protocol: "https",
+				hostname: "unavatar.io",
+			},
+		],
+	},
 };
 
 export default config;
