--- frontend/src/landing/src/app/(legal)/[slug]/page.tsx
@@ -1,5 +1,0 @@
-import { notFound } from "next/navigation";
-
-export default function NotFound() {
-  return notFound();
-}
