+++ frontend/src/landing/src/app/.well-known/api-catalog/route.ts
@@ -1,6 +1,9 @@
 import { COMPANY } from "@superset/shared/constants";
 import { API_URL, MCP_SERVER_URL } from "@/lib/llms";
 
+
+export const dynamic = "force-static";
+
 export function GET() {
   const baseUrl = COMPANY.MARKETING_URL;
