+++ frontend/src/landing/src/app/.well-known/mcp/server-card.json/route.ts
@@ -1,6 +1,9 @@
 import { COMPANY } from "@superset/shared/constants";
 import { MCP_SERVER_URL } from "@/lib/llms";
 
+
+export const dynamic = "force-static";
+
 const FALLBACK_CARD = {
   name: "ao-agents",
   description:
