--- frontend/src/landing/src/app/blog/[slug]/components/BlogPostLayout/index.ts
@@ -1,1 +1,0 @@
-export { BlogPostLayout } from "./BlogPostLayout";
