--- frontend/src/landing/src/app/changelog/[slug]/components/ChangelogEntryLayout/index.ts
@@ -1,1 +1,0 @@
-export { ChangelogEntryLayout } from "./ChangelogEntryLayout";
