+++ frontend/src/landing/src/app/index.md/route.ts
@@ -6,6 +6,8 @@ import {
   MARKDOWN_HEADERS,
 } from "@/lib/llms";
 
+export const dynamic = "force-static";
+
 export function GET() {
   const baseUrl = COMPANY.MARKETING_URL;
   const docsUrl = COMPANY.DOCS_URL;
