+++ frontend/src/landing/src/app/llms.txt/route.ts
@@ -1,5 +1,8 @@
 import { buildLlmsTxt } from "@/lib/llms";
 
+
+export const dynamic = "force-static";
+
 export async function GET() {
 	return new Response(buildLlmsTxt(), {
 		headers: {
